requirements.txt
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.3
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.8
newsapi-python==0.2.5
psycopg2==2.8.3
pycodestyle==2.5.0
pytz==2019.1
requests==2.21.0
sqlparse==0.3.0
urllib3==1.24.3
whitenoise==4.1.3
